refactor: 렌더링 최소화 하기 첫 심화 완료, 컴포넌트화
