+++ src/contexts/NotificationContext.tsx
@@ -1,0 +1,43 @@
+"use client";
+import { createContext, useContext, useState } from "react";
+import type { ReactNode } from "react";
+import type { Notification } from "../types/Notification";
+import type { NotificationContextType } from "../types/Contexts.ts";
+import { useMemo } from "../@lib";
+
+const NotificationContext = createContext<NotificationContextType | null>(null);
+
+export const NotificationProvider = ({ children }: { children: ReactNode }) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = (message: string, type: Notification["type"]) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+  };
+  const value = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications],
+  );
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+  if (!context)
+    throw new Error(
+      "useNotificationContext must be used within NotificationProvider",
+    );
+  return context;
+};
